Add svg package to latex header
Automatic conversion of SVGs to PDFs requires running pdflatex
--shell-escape to convert svgs exported by org-babel with inkscape in
the path. The converted images are cached and subsequent runs no longer
need to allow execution of external programs.

The inkscape PDF export changes the font to a serif font. For now this
is acceptable and has not been investigated in any detail.

Change-Id: I4b31483f18833679d1638bf9faade3eb19274527